pyproject.toml
[project]
name = "xplor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=8.29.0",
    "openpyxl>=3.1.5",
    "pytest>=8.3.3",
    "pytest-playwright>=0.6.2",
    "xlrd>=2.0.1",
]
